requirements.txt
paho-mqtt>=2.0,<3
stmpy>=0.7
